+++ test/helper.rb
@@ -1,45 +1,9 @@
 require 'fun_with_testing'
-
-# require 'simplecov'
-# 
-# module SimpleCov::Configuration
-#   def clean_filters
-#     @filters = []
-#   end
-# end
-# 
-# SimpleCov.configure do
-#   clean_filters
-#   load_adapter 'test_frameworks'
-# end
-# 
-# ENV["COVERAGE"] && SimpleCov.start do
-#   add_filter "/.rvm/"
-# end
-
-# require 'rubygems'
-# require 'bundler'
-
-# begin
-#   Bundler.setup(:default, :development)
-# rescue Bundler::BundlerError => e
-#   $stderr.puts e.message
-#   $stderr.puts "Run `bundle install` to install missing gems"
-#   exit e.status_code
-# end
-# require 'test/unit'
-# require 'shoulda'
-
-# $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
-# $LOAD_PATH.unshift(File.dirname(__FILE__))
-
 require 'fun_with_gems'
 
+# What exactly does this do?
 FunWith::Gems.make_gem_fun( FunWith::Testing, :require => false )
 
-# class Test::Unit::TestCase
-# end
-
 class FunWith::Gems::TestCase < FunWith::Testing::TestCase
-  include FunWith::Testing::Assertions::Basics
+  make_testing_fun( FunWith::Gems )
 end