+++ test/test_gem/lib/fun_with/llamas/gem_api.rb
@@ -1,0 +1,10 @@
+module FunWith
+  module Llamas
+    # Should be automagically 
+    module GemAPI
+      def can_make_the_llamas_dance?
+        true
+      end
+    end
+  end
+end